bikin backend

+++ database/migrations/0001_01_01_000002_create_jobs_table.php
@@ -12,16 +12,19 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('jobs', function (Blueprint $table) {
+
             $table->id();
             $table->string('queue')->index();
             $table->longText('payload');
-            $table->unsignedSmallInteger('attempts');
+            $table->unsignedTinyInteger('attempts');
             $table->unsignedInteger('reserved_at')->nullable();
             $table->unsignedInteger('available_at');
             $table->unsignedInteger('created_at');
+
         });
 
         Schema::create('job_batches', function (Blueprint $table) {
+
             $table->string('id')->primary();
             $table->string('name');
             $table->integer('total_jobs');
@@ -32,9 +35,11 @@ return new class extends Migration
             $table->integer('cancelled_at')->nullable();
             $table->integer('created_at');
             $table->integer('finished_at')->nullable();
+
         });
 
         Schema::create('failed_jobs', function (Blueprint $table) {
+
             $table->id();
             $table->string('uuid')->unique();
             $table->text('connection');
@@ -42,6 +47,7 @@ return new class extends Migration
             $table->longText('payload');
             $table->longText('exception');
             $table->timestamp('failed_at')->useCurrent();
+
         });
     }
